services/auth_service/backend/requirements.txt
Django==5.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
